Move widgets functions to the separate utils file
