+++ src/ggrc-client/js/components/assessment/info-pane/info-pane.js
@@ -54,9 +54,9 @@ import {REFRESH_TAB_CONTENT,
 } from '../../../events/eventTypes';
 import Permission from '../../../permission';
 import {
-  initCounts,
   getPageInstance,
 } from '../../../plugins/utils/current-page-utils';
+import {initCounts} from '../../../plugins/utils/widgets-utils';
 import template from './info-pane.mustache';
 import {CUSTOM_ATTRIBUTE_TYPE} from '../../../plugins/utils/custom-attribute/custom-attribute-config';
 import pubsub from '../../../pub-sub';
