+++ src/ggrc-client/js/entrypoints/export/bootstrap.js
@@ -3,7 +3,7 @@
     Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
 */
 
-import {initWidgets} from '../../plugins/utils/current-page-utils';
+import {initWidgets} from '../../plugins/utils/widgets-utils';
 import {gapiClient} from '../../plugins/ggrc-gapi-client';
 import {RouterConfig} from '../../router';
